runtime: constants and data structures for generational GC
DO NOT SUBMIT
This is the first of a CL chain introducing generational GC
to Go.

This CL introduces some high level comments about the
approach as well as the major data structures and control
variables.
The hope is that by understanding the data structures
should make the following code easily understandable.

Change-Id: I419a01322826970bd375ddea648cf8e060f8d1f1