runtime: is GC cycle full or generational
DO NOT SUBMIT
GCCycleFull determines if this GC cycle is a full
cycle. The heuristics tilt towards favoring full
cycles since that is currently the status quo. We
use some ad-hoc heuristics today. For example if
the full GC is taking less than 5% of total time
we stick with full GCs. We also don't run
generational GC for the first several GC cycles
since this is typically during initialization
when the amount of free objects is minimal.

The core heuristic is based on running a
generational GC periodically and if the amount of
time spent in a generation GC / the amount of
free memory recovered, the mark/cons ratio, is
less than with a full GC then we become more
agressive about doing generational GCs.

There is also a count based scheme included that
is turned on with a global constant that can be
used for stress testing the generational GC.

Change-Id: I9f44dbf294152df3f8ff5d87368f92fea58e37bc