runtime: make mark bits sticky
Do NOT SUBMIT
Non-moving generational GCs use the markbits from
the previous GC cycle to determine if an object
is mature. They also use the same set of mark bits
to mark reachable young objects during the current
generational GC cycle. This CL implements this logic
by using a level of indirection so that allocation
and mark bits can be one and the same during a
generational GC cycle and different when doing a
full cycle.

Change-Id: I41de3c343286bced7f5363d2858bdfc65db58f37