runtime: trigger generational GC
DO NOT SUBMIT
GC cycles are generational or full. A GC cycle
starts after mark is complete and sweeping is
about to begin. It is at this point that the
decision of whether the upcoming cycle will be
generational or a full GC is made.

If a generational GC cycle is about to begin then
the trigger for that GC is based on the live data
found in the last full GC. When half the
available allocation space is declared reachable
by a generational GC then the upcoming GC cycle
will be full.

If the previous cycle was full and the next one
will be generational the hashes need to be
invalidated. This is done by changing the base
hash seed so the upcoming hash will not match
the previous hash. This eliminates having to
redo the hashes or even clearing them.

The mark/cons ratio is calculated to help decide
if the next GC cycle will be full or
generational. A weighted decaying average (WDA)
is used to approximate the mark/cons ratio. The
WDA for full cycles with the WDA for generational
cycles are compared to decide whether the next
cycle will be full or generational.

Change-Id: Ie389a130aa156c35ff166da26416d08b519334b0