runtime: debugging and trace aids
DO NOT SUBMIT
Add a currently unused function that will
poison parts of the heap as required. This
is handy in debugging the GC since it will
trigger errors at times when a stale value
might not.

Clean up comments about when to zero objects.

Adjust some trace output to dump span state
names instead of their integer enum values.

Change-Id: I082fa0e2bc79f54edebe2413d716f6e226dca33e